+++ www/Site/Library/admin_widgets/Explorer/views/ObjectItem/ObjectItem.php
@@ -47,7 +47,7 @@ class ObjectItem extends Widget
         $v['value'] = (string)$obj->value();
         $v['uri'] = $obj->uri(false, true);
         $v['is_hidden'] = $obj->isHidden(null, false);
-        $v['is_delete'] = $obj->isDelete(null, false);
+        $v['is_draft'] = $obj->isDraft(null, false);
         $v['is_file'] = $obj->isFile();
         $v['is_link'] = $obj->isLink();
         $v['is_default_value'] = $obj->isDefaultValue();