+++ www/Site/Library/admin_widgets/Export/Export.php
@@ -89,7 +89,7 @@ class Export extends Widget
                     'select' => array('count', 'children'),
                     'from' => $obj,
                     'depth' => 'max',
-                    'where'=> array('attr', 'is_delete', '>=', 0)
+                    'where'=> array('attr', 'is_draft', '>=', 0)
                 ))+1,
                 'step' => 0,
                 'obj' => $obj->id()
@@ -120,7 +120,7 @@ class Export extends Widget
                     'select' => array('children'),
                     'from' => $info['jobs'][$j]['obj'],
                     'depth' => 'max',
-                    'where'=> array('attr', 'is_delete', '>=', 0),
+                    'where'=> array('attr', 'is_draft', '>=', 0),
                     'order'=> array(array('id', 'asc')),
                     'limit' => array($info['jobs'][$j]['step'], $cnt)
                 ));