+++ www/Site/Library/content_widgets/Page/Page.php
@@ -21,7 +21,7 @@ class Page extends AutoWidgetList2
     {
         $cond['where'] = array('all', array(
                 array('attr', 'is_hidden', '=', $this->_input['REQUEST']['object']->attr('is_hidden')),
-                array('attr', 'is_delete', '=', 0),
+                array('attr', 'is_draft', '=', 0),
                 array('attr', 'diff', '!=', Entity::DIFF_ADD)
             ));
         return parent::getList($cond);