+++ www/Site/Library/menus/Menu/item_view/item_view.php
@@ -80,7 +80,7 @@ class item_view extends AutoWidgetList2
         $cond['depth'] = array(1, 1); // выбрать из хранилища всё дерево меню
         $cond['where'] = array('all', array(
                 array('attr', 'is_hidden', '=', 0),
-                array('attr', 'is_delete', '=', 0),
+                array('attr', 'is_draft', '=', 0),
                 array('attr', 'diff', '!=', Entity::DIFF_ADD)
             ));
         //$cond['order'] = array(array('order', 'asc'));